backend/requirements.txt
# Backend Requirements
# Medical Records Bridge API Dependencies

# ============================================
# Web Framework & Extensions
# ============================================
flask==3.0.0
flask-cors==4.0.0
flask-sqlalchemy==3.1.1
flask-jwt-extended==4.5.3
gunicorn==21.2.0

# ============================================
# Database & ORM
# ============================================
sqlalchemy>=2.0.25
psycopg2-binary>=2.9.11  # PostgreSQL adapter
marshmallow==3.20.1
marshmallow-sqlalchemy==0.29.0

# ============================================
# Security & Authentication
# ============================================
bcrypt==4.1.2
python-dotenv==1.0.0
email-validator==2.1.0

# ============================================
# API Documentation
# ============================================
flasgger==0.9.7.1

# ============================================
# AI Services
# ============================================
httpx==0.25.1
numpy
huggingface_hub>=0.23.0
groq>=0.5.0

# ============================================
# OCR and Image Processing
# ============================================
pytesseract==0.3.10
Pillow==10.1.0
opencv-python==4.8.1.78

# ============================================
# Utilities
# ============================================
python-multipart==0.0.6
requests>=2.31.0  # For demo script


# Machine Learning Dependencies for OCR Training
torch>=2.1.0
torchvision>=0.16.0
transformers>=4.35.0
datasets>=2.14.0
evaluate>=0.4.0
jiwer>=3.0.0  # For CER/WER metrics

